move ThreadActiveSpanSource implementation in active_span_source module
