+++ proposal/active_span_source.py
@@ -1,3 +1,6 @@
+import threading
+
+
 class BaseActiveSpanSource(object):
     """BaseActiveSpanSource is the interface for a pluggable class that
     defines the source for the ActiveSpan. It must be used as a base class
@@ -20,3 +23,23 @@ class NoopActiveSpanSource(BaseActiveSpanSource):
 
     def active_span(self):
         return None
+
+
+class ThreadActiveSpanSource():
+    """This is a simplified implementation to make the multi-threading
+    examples work as expected. It uses a thread local storage to keep
+    track of the current ActiveSpan available in this thread execution.
+
+    Here we store the `active_span` but it could be anything else that
+    is used by Tracer developers.
+    """
+    def __init__(self):
+        self._locals = threading.local()
+
+    def make_active(self, span):
+        # implementation detail
+        setattr(self._locals, 'active_span', span)
+
+    def active_span(self):
+        # implementation detail
+        return getattr(self._locals, 'active_span', None)
