+++ proposal/helpers.py
@@ -22,23 +22,3 @@ class TracedThread(threading.Thread):
         tracer.active_span_source.make_active(self._active_span)
         del self._active_span
         super(TracedThread, self).run()
-
-
-class ThreadActiveSpanSource():
-    """This is a simplified implementation to make the multi-threading
-    examples work as expected. It uses a thread local storage to keep
-    track of the current ActiveSpan available in this thread execution.
-
-    Here we store the `active_span` but it could be anything else that
-    is used by Tracer developers.
-    """
-    def __init__(self):
-        self._locals = threading.local()
-
-    def make_active(self, span):
-        # implementation detail
-        setattr(self._locals, 'active_span', span)
-
-    def active_span(self):
-        # implementation detail
-        return getattr(self._locals, 'active_span', None)
